requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
ffmpeg-python==0.2.0
filelock==3.20.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.12.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain-core==1.1.3
langchain-nvidia-ai-endpoints==1.0.0
langsmith==0.4.57
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
narwhals==2.13.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
openai-whisper==20250625
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.2
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
streamlit==1.52.1
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
toml==0.10.2
torch==2.9.1
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
uuid_utils==0.12.0
watchdog==6.0.0
yarl==1.22.0
zstandard==0.25.0
